update: component inset and change login panel to frame
