+++ Library_System/src/view/login/IdLabel.java
@@ -7,7 +7,7 @@ import javax.swing.JLabel;
 
 public class IdLabel extends JLabel{
 	public IdLabel() {
-		String imgPath = "./image/user.png";
+		String imgPath = "./image/componentImg/user_img.png";
 		
 		ImageIcon icon = new ImageResize(imgPath,40,40).getResizeIcon();
